Logout as a distinct option, so local storage is not cleared everytime when accessing the login page

+++ copl/com.elster.jupiter.login.app/src/main/web/js/login/app/controller/Login.js
@@ -40,9 +40,13 @@ Ext.define('Login.controller.Login', {
     },
 
     showOverview: function () {
-        window.localStorage.clear();
         var params = Ext.urlDecode(location.search.substring(1));
+        if (typeof params.logout !== 'undefined') {
+            window.localStorage.clear();
+        }
+
         if (typeof params.expired !== 'undefined') {
+            window.localStorage.clear();
             var error = this.getLoginForm().down('#errorLabel');
             error.setValue('Your session has expired.');
             error.show();
